feat: leaderboard redesign, colour system pass, home screen overhaul

See merge request global-acting-in-it/goalz/lab4-goalz!80

+++ backend/Goalz/Goalz.Application/DTOs/JoinPartyRequest.cs
@@ -2,8 +2,6 @@ namespace Goalz.Core.DTOs
 {
     public class JoinPartyRequest
     {
-        public int Code {get; set;}
-        //Because there is no input field for the username, it needs to be retrieved from the JWT Token later in the Controller!
-
-    } 
+        public long Code { get; set; }
+    }
 }
